+++ src/main/java/com/sahni/converters/TaskStatusConverter.java
@@ -1,0 +1,31 @@
+package com.sahni.converters;
+
+import com.sahni.models.TaskStatus;
+
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
+import java.util.stream.Stream;
+
+@Converter(autoApply = true)
+public class TaskStatusConverter implements AttributeConverter<TaskStatus, String> {
+
+    @Override
+    public String convertToDatabaseColumn(TaskStatus taskStatus) {
+        if (taskStatus == null) {
+            return null;
+        }
+        return taskStatus.getStatus();
+    }
+
+    @Override
+    public TaskStatus convertToEntityAttribute(String taskStatus) {
+        if (taskStatus == null) {
+            return null;
+        }
+
+        return Stream.of(TaskStatus.values())
+                .filter(c -> c.getStatus().equals(taskStatus))
+                .findFirst()
+                .orElseThrow(IllegalArgumentException::new);
+    }
+}
