+++ src/main/java/com/sahni/models/TaskStatus.java
@@ -1,0 +1,17 @@
+package com.sahni.models;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum TaskStatus {
+
+    PENDING("Pending"),
+    IN_PROGRESS("In Progress"),
+    COMPLETED("Completed");
+
+    private String status;
+
+    public String getStatus() {
+        return status;
+    }
+}
